requirements.txt
opencv_python==3.4.4.19
mss==3.3.2
numpy==1.15.4
PyAutoGUI==0.9.39
Pillow==5.3.0
